+++ test/package/index.ts
@@ -1,6 +1,8 @@
 import {
   EpsonProjector,
+  ProjectorAudio,
   type EpsonProjectorOptions,
+  type ProjectorAudioOutput,
   type ProjectorPowerStatus,
 } from "easymp";
 
@@ -11,5 +13,8 @@ const options: EpsonProjectorOptions = {
 
 const projector = new EpsonProjector(options);
 const status: Promise<ProjectorPowerStatus> = projector.status.power();
+const audio: ProjectorAudio = projector.audio;
+const output: Promise<ProjectorAudioOutput> = audio.output();
 
 void status;
+void output;